Improve posting delta between two frames.
If we monitor the time between 2 frames it is not consitent.
It is expected to have posting around 33 msec for 30fps and 16 msec
for 60fps. We have an interpolation logic to take care of precise audio
time.

Logic for Interpolation:
When there is fillbuffer call presented by Audioflinger we increase the
sample count and mark the system time. In if randomly video queries for
the time then we interpolated on marked system time and use sample count
to provide a precise 1msec time line to video.

Issue with Interpolation:
We have Video posting thread and Audio time updation thread at the same
priority. Audio thread can be preempted while doing the memcpy of the
data from file to input buffer. This will communicate wrong time to
video since updation of bytes have not happened. Also in previous attempts
to read time we have used interpolation to update time. So if we have
a small byte update then we reset the interpolation marker. Then a pre-emption
for next audio time by video is even risky since we will communicate
time in past.

Solution:
To make sure we dont communicate the time in past we increase the potential
bytes we will read upfront before the thread can be pre-empted. Locking the
entire function is not a solution since memcpy delay is around 7 msec which
introduces fixed latentcy.

Defect ID: OMAPS00233908

This is a forward port of the following patch:

    commit fb25d7bee6c567c0da8a829e0bdc3e7ac2f38145
    Author: Dandawate Saket <dsaket@ti.com>
    Change-Id: I4eacc3b802a12a14155c1909bac1e56c0a05053e

Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>
Change-Id: I8ff9c80935a2cdd9f3bc803b4e9ea85713dd343a