StageFright: Smooth the audio time reference during A/V playback
The audio clock interpolation logic more or less made the following
assumptions:

* That the time when the callback AudioPlayer::fillBuffer() is
called is an accurate indication of the audio time at the end of
the buffer that it process.

* That the interpolation scheme should be updated at every one or
two calls to fillBuffer(), and unquestioningly use that
timestamp.

* That gettimeofday() is monotonic

Further, getRealTimeUs() would sometimes provide an earlier timestamp
than one in a previous call. While this strategy is one the whole
very effective, on ICS it still results in occasional frame drops at
30 fps and a little more frame jitter than is acceptable.

The strategy is modified as follows:

* The time when the callback AudioPlayer::fillBuffer() is called is
a reasonable estimation of the audio time at the end of the
PREVIOUS call to fillBuffer().

* If fillBuffer() is called multiple times quickly, the data is
aggregated and treated as though it were a single call to
fillBuffer().

* Once a timestamp is selected as the basis for the time
interpolation scheme (i.e. an epoch), we continue to use that
epoch (advancing the base timestamp based on theory instead of
measurement). We only redefine the epoch on a new timestamp if
it deviates from the audio buffer feed by more than the size of
the AudioHAL's default buffer size.

* The system monotonic clock [clock_gettime(CLOCK_MONOTONIC, ...)]
is used instead of the wall clock [gettimeofday()].

* The return value of AudioPlayer::getRealTimeUs() is made nearly
monotonic (it may repeat a previosly returned value, but never
less than a previous value).

Using this scheme, the audio time reference is very smooth and results
in less video frame jitter and fewer frame drops.

Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>
Signed-off-by: Jean Johnson <a0271255@ti.com>
Signed-off-by: Jean Johnson <a0271255@ti.com>
Signed-off-by: Devaraj Rangasamy <dev@ti.com>