To print video playback termination log when FPS automation is on
In ICS the log "video decoder shutdown completed" is currently
a LOGV. This patch changes it to LOGD if debugFps is enabled.
Similarly, it also shows frames dropped if FPS measurement is ON.

The patch is enclosed within #ifdef OMAP_ENHANCEMENT

This patch is required for automation of the video playback
regression test suite.

Redoing this patch since the previous time the patch was not
uploaded correctly to Gerrit.

Signed-off-by: a0756684@ti.com
Signed-off-by: Jean Johnson <a0271255@ti.com>
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>
Signed-off-by: Devaraj Rangasamy <dev@ti.com>