StageFright: Commit a new time interpolator.
When the previous audio clock (a.k.a. time interpolator) would get too
far ahead, it didn't have a good way to slow down.  In some use-cases
this was causing sync errors.

This patch implements a scheme that allows time-scaling to slow down
or speed up the audio clock as it tracks the hardware.  It was
inspired by the paper "Using a DLL to Filter Time" (F. Adriaensen,
2005)[1].  It differs from the paper because we are more disconnected
from the hardware, do not have constant buffer sizes, and are not
mapping frames to time-stamps.

This time source is no longer monotonic or semi-monotonic, and will
sometime "rewind" time.

[1] http://kokkinizita.linuxaudio.org/papers/usingdll.pdf

Change-Id: I43d462eb23bba19b6dd34b32b01a3c5cb7bd5d5d
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>