Stagefright: Refine audio time interpolation algorithm.
The last algorithm was focused on each call to fillBuffer().  However,
the timing of those calls was too variable to provide a smooth clock.
This patch considers the whole audio buffer (at the hardware) and
interpolates based on that.  It tracks the read pointer for the audio
media as well as the "end of the last buffer."  Using this strategy,
it scales time between now and the theoretical time when all the
current samples are played.

The expected latencies were also being determined statically.  They
are now determined dynamically using the normal Android framework.
The static values are still being used to provide sane defaults.

In practice, this strategy keeps the audio clock very steady and the
frame rate stays within +.06 / -.25 fps.  This clock is also
monotonic.

The documentation for the algorithms is in the code.

Change-Id: I1b7b163de7bc9419e038f3e7bf4eeceb66d6e7c0
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>