Stagefright: Replacing getMediaTimeUs() with getRealTimeUs()
The calculation of getMediaTimeUs() was based on number of frames played and
the timestamp present in the metadata. But with the introdudction of
interpolation logic getMediaTimeUs() is being calulated with the timestamp present
in the metadata alone.

so if the timestamps are worng then getMediaTimeUs will return wrong value which
affects in the updation of progress bar in case of audio playback as well as in the
calculation of video lag in case of AV playback.

To avoid this, it is better to use getRealTimeUs which is more accurate with the use
interpolation logic.

To fix the DR OMAPS00257698.

Change-Id: I9626c363e66fac629fccdf85ef260ac26e575a33
Signed-off-by: karimuddin sayed <sayed.karimuddin@ti.com>