Change the i/p buffer size if video clip contains bigger samples than WxH
Currently, the i/p buffer size ('nBufferSize') is always set to WxH by Ducati,
and cannot changed from Stagefright since it is a read-only paramter.
So, if video clip contains bigger samples than WxH, the video playback is failed.
This patch checks if the maximum sample size of the video clip is bigger than WxH (i.e.'nBufferSize'),
and allocate the i/p buffer with the maximum sample size if necessary.

Ref: OMAPS00259559

Change-Id: I8cc6539c6e451911f6c017276dfecb8a8a930572
Signed-off-by: Nanako Baba <baba@ti.com>
Signed-Off-By: Rajesh Vandanapu<rajeshv@ti.com>