StageFright: Update mLastVideoTimeUs and mVideoTimeUs with the timesource
The value of mVideoTimeUs is usually the value reported to userspace
for MediaPlayer.getCurrentPosition().  This value was based on the
codec (which was often updated irregularly).  Switching instead to the
AwesomePlayer's time source (usually the AudioPlayer time source) and
updating it every time a frame is posted.

This is a partial resolution to OMAPS00252206, whis is a failure in the
CTS test case android.media.cts.MediaPlayerTest#testVideoSurfaceResetting.

The test case is making pass/fail decisions based on the "current"
position of the video in the test (+/- 1.5 frames).  Because the video
timestamp is updated somewhat irregularly the test will often fail.
In contrast, the time source that AwesomePlayer uses is a more
accurate representation of the current position of the video, and the
test case will pass.

Change-Id: I2288ee38d2a9511d26ee102eb254e8d5b79bcacd
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>