AudioTrack: Do not enforce frameCount constraint.
For bluetooth audio (A2DP), AudioTrack would not create a track if the
buffer size was smaller than 1/4 the latency.  This forces BT audio to
report a large buffer size (60 ms or larger).  This causes
AudioFlinger to process audio in 60 ms chunks, which results in
unnecessary frame drops.

AudioTrack was doing this to try and ensure that writes to the HAL
will cover the latency of the device.  However, since most HAL's
allocate their own buffering (and only report a smaller "period" size
as the buffer size), this is unnecessary.

Therefore, the logic that tries to enforce this is disabled and
replaced by a warning.

This is the OMAP_ENHANCEMENT version of the following patch (which has
been submitted upstream):

    https://android-review.googlesource.com/32970
    Change-Id: I2ab61d829180b6a579629cdda7e1b1c3c443d88c

This partially resolves OMAPS00260494 "PT: sync issue with audio
through BT A2DP audio headsets on AV playback"

Change-Id: I67615d8d43b3c4804b464a43250cd6a3e818474e
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>