StageFright: Delay MEDIA_SEEK_COMPLETE until media reaches seek position
After a seek, the MEDIA_SEEK_COMPLETE signal was being sent
prematurely.  It is now being delayed until mVideoTimeUs reaches (or
exceeds) the seek position (mSeekTimeUs).

This is a partial resolution to OMAPS00252206, whis is a failure in the
CTS test case android.media.cts.MediaPlayerTest#testVideoSurfaceResetting.

This is needed for the test case, because the test tries to make a
precise measurement from a seek point to a time that is a few moments
after the seek point.  If the seek complete is sent before we actually
reach the seek point, it results in an inaccurate measurement.

Change-Id: I5e9106ae2cf9a394cf5599959b2133730ae0589c
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>