StageFright: AudioPlayer::getMediaTimeMapping() returns read pointers
When StageFright calls getMediaTimeMapping(), it is looking to compare
the codec's read pointer (mediatime_us) to the buffer-flow read
pointer (realtime_us).  However, the TimeInterpolator was implemented
with the understanding that the realtime_us should be the time
provided by the TimeInterpolator (and cached in mPositionTimeRealUs).

This is corrected by returning the TimeInterpolators read pointer
(which is m_read + m_queued).

This partially resolves OMAPS00263351 "WMV playback is jerky after
seek on OMAP4430 ICS" where this misunderstanding sometimes caused
frame drops after a seek.

Change-Id: I474e873b5ee59c6b97d938f3d1770331e4733000
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>