StageFright: Remove the "seek tolerance"
The original StageFright code was using mTimeSourceDeltaUs (based on
data returned from AudioPlayer::getMediaTimeMapping()) as an extra
compensation for A/V sync.  When the TimeInterpolator was implemented,
the seek tolerance was created because of a misunderstanding about the
function of AudioPlayer::getMediaTimeMapping().  Now that this
misunderstanding is corrected, there is no longer a need for the seek
tolerance.

This partially resolves OMAPS00263351 "WMV playback is jerky after
seek on OMAP4430 ICS."  In that DR, when the seek tolerance was met,
it altered the audio's time clock in a way that caused video jitter.

Change-Id: I09d055f10b72597027037513079ae2288f91450b
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>
Signed-off-by: Barry Woodward <b-woodward@ti.com>