StageFright: Skip the TimeInterpolator FIRST_ROLLING state.
The TimeInterpolator's FIRST_ROLLING state was a scheme to start the
AudioPlayer's clock immediately.  For the length of the latency, it
would roll along using the system's monotonic clock.  The normal
TimeInterpolator algorithm kicked in after the latency had elapsed.
When the TimeInterpolator algorithm engaged, it would typically slow
down the clock for a short period (about .5 sec) until the clock
aligned with the start-up latency.  (The start-up latency is usually
not the same as the latency.)

The result is that the first 1-2 secs of video aren't synchronized as
well as the rest of the video.

This patch bypasses the FIRST_ROLLING state.  Suppose the video time
is at 0.  When the video starts rolling, m_pos0 is moved from 0 to
-latency/2.  For a while, the TimeInterpolator provides negative
timestamps.

In practice, good sync is typically achieved within 3-4 video frames.
This is partly due to using latency/2 as an estimate of the start-up
latency.

This is an intermediate patch.  The following patch will totally
remove the FIRST_ROLLING and FIRST_PAUSED states.

Change-Id: Iea3f3c15efa17e7f2df9418645b2c98f42abb167
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>