Workaround for Mpeg2TS grey thumbnail issue
Seek is not supported in Mpeg2TSExtractor. Thus during thumbnail
generation if the software decoder is failing the extractor is not
seeking to the I-frame and is providing the next available frame to
hardware decoder. Thus there is issue in thumbnail generation.
To resolve this, we directly use hardware decoder for all the
Mpeg2Ts Clips.

OMAP_ENHANCEMENT macro has been added. It is a workaround,
and the actual fix should be done inside Mpeg2TSParser.
Changes are added for Mpeg2PS clips as well. Now for Mpeg2TS,
Mpeg2PS clips by default hardware codec is used for thumbnail
generation.

Change-Id: I89c4b31b108834b59d4264204ef270cf22c8ab47
Signed-off-by: Santhosh Behara <santhosh.behara@ti.com>