Limit maximum wallpaper size
This patch addresses the issue of wallpapers
using textures too large for SGX to handle.
In the case of SGX540, a texture cannot be larger
than 2048x2048 pixels. The Launcher2 app needs
some extra width for wallpaper parallax effect,
and when the framebuffer is set to 1080p the
requested wallpaper size goes beyond the SGX limit.
As a result, the wallpaper manager service fails and
SystemUI app crashes.

Change-Id: Ic95fc29a1287b8320fcb3eae2d646b1b27f87884
Signed-off-by: Tate Hornbeck <tate.hornbeck@ti.com>