TimeLapse recording with High Profile Encoder
When Timelapse recording is performed, Encoder is expected
to process each input buffer and provide a output buffer.
This wont work when we have B Frames enabled.
This patch disables B Frames when Encoder flag
kOnlySubmitOneInputBufferAtOneTime is set to allow
submission of 1 frame at a time.

Patch set#4: Added the check to cover all encoders.
Updated commit message and comments in code to be in sync
with the patch submitted to google.

Change-Id: I7dd17f8ce9185abe3c6749caa34c90500bd2f913
Signed-off-by: Sathya Prakash M R <a0393677@ti.com>