AwesomePlayer always reads the forward syncable video data after SEEK.
By the patch#19882(http://review.omapzoom.org/#/c/19882/), the SEEK_COMPLETE
event happens after mVideoTimeUs reaches to the seek position.
This means that video frames continue to come from the "old" video position,
and only after MEDIA_SEEK_COMPLETE signal is sent, video frames come from
the new position. So, if the new position is the backward of the seek position,
it is sometimes confusing because the video player seems to ignore (briefly)
the user's input.
With the patch, the new position is always forward of the seek position, and
MEDIA_SEEK_COMPLETE signal is sent immediately. So, video frame from
the new position will be displayed immediately, too.
This gives the user positive visual feedback that their input has been accepted.

Change-Id: I72327ba08f060a5b474584c6084875ec02afebca
Signed-off-by: Nanako Baba <baba@ti.com>