MPEG4Extractor: Handle the haveDeltaTable() function via sub-classing
Change Ib1dbdf0158a93616dc917bfe031ef62597c4b0d5 added a virtual
function to MediaSource which breaks binary compatibility with 3rd
party binaries.  For example, this broke WMV playback because the
vtables didn't match.

This is corrected creating a subclass of MediaSource called
MediaSourceHaveDeltaTable and deriving MPEG4Source from the new
subclass.  In AwesomePlayer, if the container is MPEG4 then we
statically cast to MediaSourceHaveDeltaTable to use the new function.

Change-Id: I36867aaa0a61050efef4fdf9bc840b72c3b5b7d0
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>