telephony: neighboring cells are now handled properly
GsmServiceStateTracker.java sets the System Property
PROPERTY_DATA_NETWORK_TYPE = "TypeName:RadioTechnology"
for the current network type (ex. EDGE:2 or HSPA:11).

When neighboring cell info is sent to responseCellList() in RIL.java,
the routine compares PROPERTY_DATA_NETWORK_TYPE to just the TypeName
(without the :) and can not find a match.  It defaults to
NETWORK_TYPE_UNKNOWN and does no interpret the location.

This patch splits off the string part of PROPERTY_DATA_NETWORK_TYPE
so a proper match is found and the location is interpreted.

Also HSPA+ (HSPAP) is added as a valid match.

Patch compiled and tested on HP TouchPad 4G.

Change-Id: I223db80d6c4a7791300e299826aa316854738d5c

Conflicts:

	telephony/java/com/android/internal/telephony/RIL.java